requirements.txt
# Runtime dependencies
pyyaml>=6.0
groq>=0.9.0
requests>=2.31.0
ruff>=0.1.0

# Optional AI providers (install the one you use)
# anthropic>=0.25.0
# openai>=1.0.0
# google-genai>=1.0.0

# Development / testing
pytest>=7.4
pytest-cov>=4.1
